// shared/src/lib/schemas.ts
export const notesRowSchema = z.object({
  context: z.string().nullable(),
  context_html: z.string().nullable(),
  created_at: z.string(),
  highlights: z.array(z.string()),
  id: z.number(),
  predicted_topic: z.string().nullable(),
  quote: z.string(),
  serialized_highlight: z.string().nullable(),
  snippet_uuid: z.string().nullable(),
  source_id: z.number(),
  tags: z.array(z.string()),
  user_id: z.string(),
  user_note: z.string().nullable(),
})